Refactor bigval_t struct and move it to gc-common.h
